feat: extract line logic into new service, fix
This commit ensures that _if_ a line is marked as `isEol` true _but_ the
status is somehow not `EOL`, an error is thrown in the ui.

Additionally, this commit lays the groundwork for more robust parsing.

The eol scanner command (eol.ts), returns ScanResultComponent[] from an
sbom.
Then, the `eol.svc` service transforms ScanResultComponents into Lines.

Something like zod would be useful to actually ensure that we are
parsing ScanResultComponents into lines in a type-safe manner.

This commit starts us down that road by extracting Line-related logic
into a new service.