chore: add warning for unknown purls
This commit adds a warning in cases where a generated sbom has a purl,
but the NES/XEOL database does not have any information.

We may want to fail silently instead. But even if we don't want the CLI
to report anything to the user, it may be worth it to eventually
track EOL 'misses' in some form.

This commit also improves typing in several ways that makes
it easier to trace why scan details might be missing in the first place:

- Use an actual Map when building scan results
- set ScanResultComponent.status to optional to match current api
- create ComponentStatus type union
- replace optional Line properties with defined properties where feasible