feat: refactor extract-components to just purls
In some cases, cyclonedx attaches metadata in the sbom about where
a purl occurred in the codebase: sbom.components.evidence.occurrences.

Not every component has occurrences.

The CLI currently does not do anything with occurrence data, and it is
not part of the v0 or v1 requirements to display it to the user.

If we can delete occurrences from the final report ui, we can greatly
simplify the data flow into a pipeable set of operations:

=> generate sbom (or load pre-existing sbom)
=> get purls from sbom
=> post purls to monorepo api
=> transform api response (scan) to user report
=> download report and/or send report to herodevs