feat: add --save and --file flags (#94)
* feat: refactor extract-components to just purls

In some cases, cyclonedx attaches metadata in the sbom about where
a purl occurred in the codebase: sbom.components.evidence.occurrences.

Not every component has occurrences.

The CLI currently does not do anything with occurrence data, and it is
not part of the v0 or v1 requirements to display it to the user.

If we can delete occurrences from the final report ui, we can greatly
simplify the data flow into a pipeable set of operations:

=> generate sbom (or load pre-existing sbom)
=> get purls from sbom
=> post purls to monorepo api
=> transform api response (scan) to user report
=> download report and/or send report to herodevs

* chore: alphabetize scripts, add readme script

* chore: add cspell words

* fix: update method that checks for eol status

Only throw error if component status is set to a truthy value.

* feat: add --save and --file flags

The save flag allows a user to save a generated sbom to the project.

The file flag allows a user to load a pre-existing cyclonedx
sbom and use that for the eol scan. Skipping the cli step of
generating a sbom.

* chore: update readme with new flags

* feat: create purls report command

This command generates a report of purls for the user.

* feat: simplify extract-purls method

Input is always valid since its validated by validateIsCycloneDxSbom

* chore: refactor error handling in save method

Duck typing > instance checking