feat: implement report committers command (#106)
* feat: implement report committers command

The new `report committers` command can be used to
generate a report of recent committers to a project.

By default, it will look for contributions over the past 12 months.

The time period can be customized using the `--months` flag.

The command optionally saves the output to file. The output can
be set to text, csv, or json format.

Please note, this command leverages `git log` built-in methods
for handling commits since a certain date. To the extent `git log`
fails to handle edge cases such as Daylight Savings time and/or
timezone issues, this command will also fail to handle those scenarios.

If a more accurate date-parsing feature is necessary it can be
revisited when needed.

* feat: ensure all committers to repo are counted

Including committers on other branches.

* chore: limit release step to workflow_dispatch

* feat: use modern node syntax for object handling