feat: add purls flag to eol scan (#141)
This flag allows eol scan to directly consume a list of purls to send to
our api.

Users can still generate an sbom or use an existing sbom to get a list
of purls.

This feature allows `scan eol` to be useable by clients who want
to derive a list of purls using a method besides cyclonedx sboms.

This feature also allows `scan eol` to `dog food` the output of
`report purls -s`.