feat!: implements new pagination api (#149)
* feat: introduce scan input options (wip)

Decouple cli interfaces from backend graphql types.

* feat: add insights eol scan input interface

This is the interface from the backend.

* feat!: use pagination for sending large scans

The backend is introducing new required fields `page` and `totalPage`.

A new deployment is necessary on the cli side and the
backend to accommodate this breaking api change.

BREAKING CHANGE: introduce pagination

* feat: refactor batch processing

The backend api has changed for how batches are processed.

If there are multiple pages, then the backend sends a placeholder
response for each set of purls until the final set has been received.

When the backend has finished processing the final result, it will send
a scan report.

* chore: move submitScan to top of file

* chore: release 1.3.0-beta.1

Release-As: 1.3.0-beta.1

BREAKING CHANGE: implements new pagination api (#149)