feat: add update-notifier package
Users who install the package using `npm i -g ...` or even `npx @...`
should be notified when there is a more recent version of the CLI
that is available.

The npm package `update-notifier` satisfies this feature.

For users that install the tarball directly from s3, oclif will handle
automatically updating the package for them.