fix: properly display package names on table (#183)
* fix: properly display package names on table

Previously, when running `npm run dev -- scan eol -t` the data in the
name column was incorrect for packages with a namespace.

For example, the name for @angular/core was displayed only as `core`.

This commit ensures that namespaces are properly combined with names
for a more accurate package name display in table output.

* chore: add specs for purl package name resolver