fix: ensure e2e tests don't run against prod (#250)
* fix: ensure e2e tests don't run against prod

Previously, the e2e tests were being run against prod. This was not the
intention. Process env variables were set in multiple places:
- bin/run.js
- bin/dev.js
- eol.test.ts
- src/config/constants.ts

This allowed for the `config` and the `process.env` vars to get out
of sync, which caused unexpected results for the value of those vars
during e2e tests.

This commit consolidates to the `config` as the source of truth for
env vars. I've deleted the attempts to set env vars everywhere else.

For convenience, I've added a `.envrc.example` file showing how to set
development vars. You can use `direnv allow` to load vars in a .envrc
file.[1]

[1]https://github.com/direnv/direnv/tree/master

* fix: move env vars into correct part of test step

* chore: temporarily skip x-large purls spec